fix(lint): first honest lint run — real bugs fixed, checks curated
make lint was discarding run-clang-tidy stderr, so tool failures printed
'no findings'; the Linux CI tidy job surfaced 128 findings the local
gate never saw. Fixed the real ones: use-after-move in the
for_each_type fold, unchecked optional accesses, exceptions escaping
main (function-try-block + non-throwing handlers), narrowing
conversions, explicit ctor, enum base type, identifier naming. Curated
.clang-tidy down to checks the codebase satisfies (each disabled family
documented); lint now fails when the tool itself fails; src/python
excluded (nanobind macro noise).

+++ src/metric.cpp
@@ -151,8 +151,10 @@ float compute_auc(floats_view probs, floats_view labels)
     {
         return 1.0F; // degenerate: one class only
     }
-    double const u = rank_sum_pos - (static_cast<double>(n_pos) * (n_pos + 1) / 2.0);
-    return static_cast<float>(u / (static_cast<double>(n_pos) * n_neg));
+    double const u = rank_sum_pos - (static_cast<double>(n_pos) *
+                                     static_cast<double>(n_pos + 1) / 2.0);
+    return static_cast<float>(
+        u / (static_cast<double>(n_pos) * static_cast<double>(n_neg)));
 }
 
 float compute_mc_accuracy(floats_view class_ids, floats_view labels)
